Submitting the payment

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -16,21 +16,21 @@ declare var Stripe;
 })
 export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   @Input() checkoutForm: FormGroup;
-  @ViewChild('cardNumber',{ static: true}) cardNumberElement: ElementRef;
-  @ViewChild('cardExpiry',{ static: true}) cardExpiryElement: ElementRef;
-  @ViewChild('cardCvc',{ static: true}) cardCvcElement: ElementRef;
-  stripe:any;
-  cardNumber:any;
+  @ViewChild('cardNumber', { static: true }) cardNumberElement: ElementRef;
+  @ViewChild('cardExpiry', { static: true }) cardExpiryElement: ElementRef;
+  @ViewChild('cardCvc', { static: true }) cardCvcElement: ElementRef;
+  stripe: any;
+  cardNumber: any;
   cardExpiry: any;
   cardCvc: any;
   cardErrors: any;
   cardHandler = this.onChange.bind(this);
 
   constructor(
-    private basketService:BasketService, 
-    private checkoutService:CheckoutService,
-    private toastr:ToastrService,
-    private router:Router) { }
+    private basketService: BasketService,
+    private checkoutService: CheckoutService,
+    private toastr: ToastrService,
+    private router: Router) { }
 
   ngAfterViewInit() {
     this.stripe = Stripe('pk_test_51JkSzKFH6IHDh4o0WNUcMVf5Nei3qCFCg2BvgwpjxzA9d69ZuZ2F4JCE5DVScgH87opBRtWIkfbDNLWqA4KAIjxz00rU5vGqKd');
@@ -39,11 +39,11 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
     this.cardNumber = elements.create('cardNumber');
     this.cardNumber.mount(this.cardNumberElement.nativeElement);
     this.cardNumber.addEventListener('change', this.cardHandler);
-    
+
     this.cardExpiry = elements.create('cardExpiry');
     this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
     this.cardExpiry.addEventListener('change', this.cardHandler);
-    
+
     this.cardCvc = elements.create('cardCvc');
     this.cardCvc.mount(this.cardCvcElement.nativeElement);
     this.cardCvc.addEventListener('change', this.cardHandler);
@@ -55,23 +55,37 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
     this.cardCvc.destroy();
   }
 
-  onChange({error}){
-    if(error) {
+  onChange({ error }) {
+    if (error) {
       this.cardErrors = error.message;
-    }else{
+    } else {
       this.cardErrors = null;
     }
   }
 
-  submitOrder(){
+  submitOrder() {
     const basket = this.basketService.getCurrentBasketValue();
     const orderToCreate = this.getOrderToCreate(basket);
-    this.checkoutService.createOrder(orderToCreate).subscribe((order:IOrder) =>{
+    this.checkoutService.createOrder(orderToCreate).subscribe((order: IOrder) => {
       this.toastr.success('Order created successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const navigationExtras:NavigationExtras = {state:order};
-      this.router.navigate(['checkout/success'], navigationExtras);
-    },error =>{
+      this.stripe.confirmCardPayment(basket.clientSecret, {
+        payment_method: {
+          card: this.cardNumber,
+          billing_details: {
+            name: this.checkoutForm.get('paymentForm').get('nameOnCard').value
+          }
+        }
+      }).then(result => {
+        console.log(result);
+        if (result.paymentIntent) {
+          this.basketService.deleteLocalBasket(basket.id);
+          const navigationExtras: NavigationExtras = { state: order };
+          this.router.navigate(['checkout/success'], navigationExtras);
+        } else {
+          this.toastr.error('payment error');
+        }
+      });
+    }, error => {
       this.toastr.error(error.message);
       console.log(error);
     })
